requirements.txt
web3>=6.0.0
click
python-dotenv
tqdm
